+++ common/src/main/java/dev/ftb/mods/ftbultimine/client/CachedEdge.java
@@ -1,8 +1,5 @@
 package dev.ftb.mods.ftbultimine.client;
 
-/**
- * @author LatvianModder
- */
 public class CachedEdge {
 	public float x1, y1, z1, x2, y2, z2;
 }