+++ common/src/main/java/dev/ftb/mods/ftbultimine/client/FTBUltimineClient.java
@@ -12,6 +12,7 @@ import dev.architectury.event.events.client.ClientRawInputEvent;
 import dev.architectury.event.events.client.ClientTickEvent;
 import dev.architectury.registry.client.keymappings.KeyMappingRegistry;
 import dev.ftb.mods.ftblibrary.config.ui.EditConfigScreen;
+import dev.ftb.mods.ftbultimine.CooldownTracker;
 import dev.ftb.mods.ftbultimine.FTBUltimine;
 import dev.ftb.mods.ftbultimine.FTBUltimineCommon;
 import dev.ftb.mods.ftbultimine.config.FTBUltimineClientConfig;
@@ -32,19 +33,16 @@ import net.minecraft.network.chat.Component;
 import net.minecraft.network.chat.MutableComponent;
 import net.minecraft.network.chat.TextColor;
 import net.minecraft.util.FormattedCharSequence;
+import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.phys.AABB;
 import net.minecraft.world.phys.Vec3;
 import net.minecraft.world.phys.shapes.BooleanOp;
 import net.minecraft.world.phys.shapes.Shapes;
 import net.minecraft.world.phys.shapes.VoxelShape;
 import org.joml.Matrix4f;
-import org.lwjgl.glfw.GLFW;
 
 import java.util.*;
 
-/**
- * @author LatvianModder
- */
 public class FTBUltimineClient extends FTBUltimineCommon {
 	public static KeyMapping keyBinding;
 	private boolean pressed;
@@ -59,7 +57,7 @@ public class FTBUltimineClient extends FTBUltimineCommon {
 	private int shapeIdx = 0;  // shape index of client player's current shape
 
 	public FTBUltimineClient() {
-		KeyMappingRegistry.register(keyBinding = new KeyMapping("key.ftbultimine", InputConstants.Type.KEYSYM, GLFW.GLFW_KEY_GRAVE_ACCENT, "key.categories.ftbultimine"));
+		KeyMappingRegistry.register(keyBinding = new KeyMapping("key.ftbultimine", InputConstants.Type.KEYSYM, InputConstants.KEY_GRAVE, "key.categories.ftbultimine"));
 
 		ClientLifecycleEvent.CLIENT_LEVEL_LOAD.register(__ -> FTBUltimineClientConfig.load());
 		ClientLifecycleEvent.CLIENT_SETUP.register(this::onClientSetup);
@@ -70,8 +68,12 @@ public class FTBUltimineClient extends FTBUltimineCommon {
 		// TODO: remove once #6 gets fixed
 		LevelRenderLastEvent.EVENT.register(this::renderInGame);
 
-		ClientRawInputEvent.MOUSE_SCROLLED.register(this::mouseEvent);
+		ClientRawInputEvent.MOUSE_SCROLLED.register(this::onMouseScrolled);
 		ClientRawInputEvent.KEY_PRESSED.register(this::onKeyPress);
+	}
+
+	public static Player getClientPlayer() {
+		return Minecraft.getInstance().player;
 	}
 
 	private void onClientSetup(Minecraft minecraft) {
@@ -91,9 +93,9 @@ public class FTBUltimineClient extends FTBUltimineCommon {
 	@Override
 	public void editConfig(boolean isClientConfig) {
 		if (isClientConfig) {
-			new EditConfigScreen(FTBUltimineClientConfig.getConfigGroup()).openGui();
+			new EditConfigScreen(FTBUltimineClientConfig.getConfigGroup()).setAutoclose(true).openGui();
 		} else {
-			new EditConfigScreen(FTBUltimineServerConfig.getConfigGroup()).openGui();
+			new EditConfigScreen(FTBUltimineServerConfig.getConfigGroup()).setAutoclose(true).openGui();
 		}
 	}
 
@@ -129,8 +131,8 @@ public class FTBUltimineClient extends FTBUltimineCommon {
 		VertexConsumer buffer2 = mc.renderBuffers().bufferSource().getBuffer(UltimineRenderTypes.LINES_TRANSPARENT);
 
 		for (CachedEdge edge : cachedEdges) {
-			buffer2.vertex(matrix, edge.x1, edge.y1, edge.z1).color(255, 255, 255, 10).endVertex();
-			buffer2.vertex(matrix, edge.x2, edge.y2, edge.z2).color(255, 255, 255, 10).endVertex();
+			buffer2.vertex(matrix, edge.x1, edge.y1, edge.z1).color(255, 255, 255, 30).endVertex();
+			buffer2.vertex(matrix, edge.x2, edge.y2, edge.z2).color(255, 255, 255, 30).endVertex();
 		}
 
 		mc.renderBuffers().bufferSource().endBatch(UltimineRenderTypes.LINES_TRANSPARENT);
@@ -138,10 +140,10 @@ public class FTBUltimineClient extends FTBUltimineCommon {
 		stack.popPose();
 	}
 
-	public EventResult mouseEvent(Minecraft client, double amount) {
-		if (pressed && amount != 0 && sneak()) {
+	public EventResult onMouseScrolled(Minecraft client, double amountX, double amountY) {
+		if (pressed && (amountY != 0 || amountX != 0) && sneak()) {
 			hasScrolled = true;
-			new ModeChangedPacket(amount < 0D).sendToServer();
+			new ModeChangedPacket(amountX < 0D || amountY < 0D).sendToServer();
 			return EventResult.interruptFalse();
 		}
 
@@ -153,7 +155,7 @@ public class FTBUltimineClient extends FTBUltimineCommon {
 			return EventResult.pass();
 		}
 
-		if (keyCode != GLFW.GLFW_KEY_UP && keyCode != GLFW.GLFW_KEY_DOWN) {
+		if (keyCode != InputConstants.KEY_UP && keyCode != InputConstants.KEY_DOWN) {
 			return EventResult.pass();
 		}
 
@@ -162,7 +164,7 @@ public class FTBUltimineClient extends FTBUltimineCommon {
 		}
 
 		hasScrolled = true;
-		new ModeChangedPacket(keyCode == GLFW.GLFW_KEY_DOWN).sendToServer();
+		new ModeChangedPacket(keyCode == InputConstants.KEY_DOWN).sendToServer();
 		lastToggle = System.currentTimeMillis();
 		return EventResult.pass();
 	}
@@ -170,18 +172,26 @@ public class FTBUltimineClient extends FTBUltimineCommon {
 	private boolean sneak() {
 		if (!FTBUltimineClientConfig.requireSneakForMenu.get()) return true;
 
-		return keyBinding.key.getValue() == GLFW.GLFW_KEY_LEFT_SHIFT || keyBinding.key.getValue() == GLFW.GLFW_KEY_RIGHT_SHIFT ? Screen.hasControlDown() : Screen.hasShiftDown();
+		return keyBinding.matches(InputConstants.KEY_LSHIFT, 0) || keyBinding.matches(InputConstants.KEY_RSHIFT, 0) ?
+				Screen.hasControlDown() :
+				Screen.hasShiftDown();
 	}
 
 	private void addPressedInfo(List<MutableComponent> list) {
-		list.add(Component.translatable("ftbultimine.info.base",
-				canUltimine && actualBlocks > 0 ?
-						Component.translatable("ftbultimine.info.active").withStyle(style -> style.withColor(TextColor.fromRgb(0xA3BE8C))) :
-						Component.translatable("ftbultimine.info.not_active").withStyle(style -> style.withColor(TextColor.fromRgb(0xBF616A)))
-		));
+		Component msg;
+		if (CooldownTracker.isOnCooldown(getClientPlayer())) {
+			msg = Component.translatable("ftbultimine.info.cooldown").withStyle(style -> style.withColor(TextColor.fromRgb(0xBFBF8C)));
+		} else if (canUltimine && actualBlocks > 0) {
+			msg = Component.translatable("ftbultimine.info.active").withStyle(style -> style.withColor(TextColor.fromRgb(0xA3BE8C)));
+		} else {
+			msg = Component.translatable("ftbultimine.info.not_active").withStyle(style -> style.withColor(TextColor.fromRgb(0xBF616A)));
+		}
+		list.add(Component.translatable("ftbultimine.info.base", msg));
 
 		if (!hasScrolled) {
-			list.add(Component.translatable("ftbultimine.change_shape").withStyle(ChatFormatting.GRAY));
+			MutableComponent msg1 = Component.translatable(FTBUltimineClientConfig.requireSneakForMenu.get() ?
+					"ftbultimine.change_shape" : "ftbultimine.change_shape.no_shift").withStyle(ChatFormatting.GRAY);
+			list.add(msg1);
 		}
 
 		int context = Math.min((ShapeRegistry.shapeCount() - 1) / 2, FTBUltimineClientConfig.shapeMenuContextLines.get());
@@ -242,12 +252,22 @@ public class FTBUltimineClient extends FTBUltimineCommon {
 			Minecraft minecraft = Minecraft.getInstance();
 
 			int top = 2 + minecraft.font.lineHeight * infoOffset;
-
+			boolean first = true;
 			for (MutableComponent msg : list) {
 				FormattedCharSequence formatted = msg.getVisualOrderText();
-				graphics.fill(1, top - 1, 2 + minecraft.font.width(formatted) + 1, top + minecraft.font.lineHeight - 1, 0xAA_2E3440);
+				if (first) {
+					float f = CooldownTracker.getCooldownRemaining(getClientPlayer());
+					if (f < 1f) {
+						graphics.fill(1, top - 1, 2 + (int)(minecraft.font.width(formatted) * f) + 1, top + minecraft.font.lineHeight - 1, 0xAA_2E3440);
+					} else {
+						graphics.fill(1, top - 1, 2 + minecraft.font.width(formatted) + 1, top + minecraft.font.lineHeight - 1, 0xAA_2E3440);
+					}
+				} else {
+					graphics.fill(1, top - 1, 2 + minecraft.font.width(formatted) + 1, top + minecraft.font.lineHeight - 1, 0xAA_2E3440);
+				}
 				graphics.drawString(minecraft.font, formatted, 2, top, 0xECEFF4, true);
 				top += minecraft.font.lineHeight;
+				first = false;
 			}
 		}
 	}