+++ common/src/main/java/dev/ftb/mods/ftbultimine/client/ShapeScreen.java
@@ -1,7 +1,4 @@
 package dev.ftb.mods.ftbultimine.client;
 
-/**
- * @author LatvianModder
- */
 public class ShapeScreen {
 }