+++ common/src/main/java/dev/ftb/mods/ftbultimine/net/FTBUltimineNet.java
@@ -4,9 +4,6 @@ import dev.architectury.networking.simple.MessageType;
 import dev.architectury.networking.simple.SimpleNetworkManager;
 import dev.ftb.mods.ftbultimine.FTBUltimine;
 
-/**
- * @author LatvianModder
- */
 public interface FTBUltimineNet {
 	SimpleNetworkManager NET = SimpleNetworkManager.create(FTBUltimine.MOD_ID);
 
@@ -16,6 +13,7 @@ public interface FTBUltimineNet {
 	MessageType SYNC_CONFIG_FROM_SERVER = NET.registerS2C("sync_config_from_server", SyncConfigFromServerPacket::new);
 	MessageType SYNC_CONFIG_TO_SERVER = NET.registerC2S("sync_config_to_server", SyncConfigToServerPacket::new);
 	MessageType EDIT_CONFIG = NET.registerS2C("edit_config", EditConfigPacket::new);
+	MessageType SYNC_ULTIMINE_TIME = NET.registerS2C("sync_ultimine_time", SyncUltimineTimePacket::new);
 
 	static void init() {
 	}