+++ common/src/main/java/dev/ftb/mods/ftbultimine/shape/EscapeTunnelShape.java
@@ -1,13 +1,5 @@
 package dev.ftb.mods.ftbultimine.shape;
 
-import net.minecraft.core.BlockPos;
-
-import java.util.ArrayList;
-import java.util.List;
-
-/**
- * @author LatvianModder
- */
 public class EscapeTunnelShape extends DiagonalTunnelShape {
 	@Override
 	public String getName() {