+++ common/src/main/java/dev/ftb/mods/ftbultimine/shape/ShapeContext.java
@@ -6,9 +6,6 @@ import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.state.BlockState;
 
-/**
- * @author LatvianModder
- */
 public record ShapeContext(ServerPlayer player, BlockPos pos, Direction face, BlockState original, BlockMatcher matcher, int maxBlocks) {
 	public boolean check(BlockState state) {
 		return matcher.actualCheck(original, state);