+++ common/src/main/java/dev/ftb/mods/ftbultimine/shape/SmallSquareShape.java
@@ -1,8 +1,5 @@
 package dev.ftb.mods.ftbultimine.shape;
 
-/**
- * @author LatvianModder
- */
 public class SmallSquareShape extends LargeTunnelShape {
 	@Override
 	public String getName() {