+++ neoforge/src/main/java/dev/ftb/mods/ftbultimine/forge/plugin/losttrinkets/LostTrinketsFTBUltiminePlugin.java
@@ -1,0 +1,27 @@
+package dev.ftb.mods.ftbultimine.forge.plugin.losttrinkets;
+
+import dev.ftb.mods.ftbultimine.integration.FTBUltiminePlugin;
+
+public class LostTrinketsFTBUltiminePlugin implements FTBUltiminePlugin {
+
+//	private static final DeferredRegister<Item> ITEMS = DeferredRegister.create(Item.class, FTBUltimine.MOD_ID);
+
+//	private static RegistryObject<UltiminerTrinket> TRINKET;
+
+//	public LostTrinketsFTBUltiminePlugin() {
+//		ITEMS.register(FMLJavaModLoadingContext.get().getModEventBus());
+//	}
+
+//	@Override
+//	public void init() {
+//		if (!FTBUltimineCommonConfig.USE_TRINKET.get()) {
+//			return;
+//		}
+//		TRINKET = ITEMS.register("ultiminer", UltiminerTrinket::new);
+//	}
+
+//	@Override
+//	public boolean canUltimine(Player player) {
+//		return !FTBUltimineCommonConfig.USE_TRINKET.get() || LostTrinketsAPI.getTrinkets(player).isActive(TRINKET.get());
+//	}
+}