+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2004.1.0]
+
+### Changed
+* Ported to Minecraft 1.20.4. Supported on Forge, NeoForge and Fabric.
+
+## [2001.1.3]
+
+### Added
+* Added new block tag `ftbultimine:block_whitelist`
+  * If this tag is non-empty, then _only_ blocks in this tag can be ultimined
+  * Complements the existing `ftbultimine:exclude_blocks` tag which can be used to blacklist blocks
 
 ## [2001.1.2]